Added a callback_manager.h to handle lists of callbacks.
Added a test to verify that the callback_manager works as expected.
Added the is_smart_pointer.h here so that way we can use it anywhere.
Added missing is_string_literal.h to the install() rule.
Removed the PROJECT_TEST_ARGS of the --tmp, we now have a --tmp-dir in snapcatch2.
Fixed the corresponding uses of g_tmp_dir, which is now a function.
Removed support for the --tmp command line option.
Cleaned up the copyright notice in the main CMakeLists.txt file.