Changed the mechanism to add/remove callbacks.
In order to be able to remove any callback, the easiest is to generate a
unique identifier for each callback. We use a 32 bit integer for that purpose.
That id is returned by the add_callback() and can be reused in the
remove_callback() to remove that given callback from the manager.

I completely removed the DUPLICATES idea. It's not useful and it wouldn't
work well with objects such as std::bind() which do not offer an == operator.
Also, it should be easy enough for your code to not re-add the same callback
over and over again.