Added a SAFE_ASSERT() function.
The assert() "function" is notoriously unsafe because it uses a macro. Now a
day, there are features in each compiler to verify for side effects, but the
truth is that if you call a function that gets completely removed at runtime,
you may end up with a side effect issue between Debug and Release.

Added a test to verify the SAFE_ASSERT().