Created a timespec_ex class.
Converted the timespec "bare functions" in a class which allows us to get
the operators even though the class is defined inside a namespace.

Added a negative() function to test whether the value is considered negative.

Added gettime() to setup the timespec_ex to the system time.

Simplified the operators since the constructors allow for comparing with
timespec, int64_t and double without the need to create one operator per
type. However, we still can't have a bool cast operator.