Added the saturated_add() functions for integers.
Fixed power128() for speed and made it available for all integral types.
Also fixed pow[er128]() for negative powers.
Fixed the std::numerical_limits for int128 numbers.
Fixed the mask used in brs as the size is actually 23 bits.
Fixed the int128 functions with inlines since these are in headers.
Fixed in brs test that could generate duplicate keys.