Fixed two bugs with the recursive functionality.
The "..." pattern has to be checked before we check the f_recursive flag.
If f_recursive is true and the directory does not exist, we would return true.
That was incorrect. We had to return false if the directory cannot be found.
You can distinguish that error with most others using
get_last_error_errno() == ENOENT (which ipload.cpp does at the moment).