Added support for checking versions.
This version includes two new macros which are very useful to properly test
that a library is at the right version to be used as compile time.

Also make use of that in the catch_rm_r.cpp since we need to do so to start
using the compiler on Ubuntu 22.04.