Added the "safe_object" template.
"Often", I allocate a resource and need to keep it in a non-shared pointer
type of storage. However, until the full initialization is done, the resource
is at risk. This object allows me to minimize the risk and still be able
to deallocate if an exception occur between the allocation and the
finalization of the initialization process.

I updated the test tags to something a bit more useful.