Removing the snapdev-doc.
I had that only because I wanted to not have a special case in the build
system but now I see that the "install" file was incorrect either way.
The newer versions of debhelper are way better since they detect way more
errors that you can then fix and make your packages better.