Added NOT_REACHED_IN_TEST() for added protection.
Many of my tests cannot reach certain parts of the code, or at least, I
suspect that it's not reachable. If it does get there, though, we want
to get an error when running the final coverage test. This function
does nothing when not running those tests. However, if running those
tests and it gets called, then it abort()'s the test, which the catch2
library captures. As a result, we can be sure that it really wasn't
called. The function cames with an example on how to use it.