Adding a "float to string" function.
The function was taken from as2js (the floating_point.h). That way we can
convert floating points to a string in other places too. At the moment
it would be useful in prinbee. This function converts NaN and Infinity
as such strings instead of failing or returning an invalid floating
point representation.