Fixed one TODO with an std::clog message.
In the file_contents, the deletion of the temporary file may fail in which
case we'd like to get a message. By default, it will be a message in stderr.
If the command runs with snaplogger, then it is captured and sent to the
logger output (as per its appenders).

Also enhanced the SAFE_ASSERT() docs.