Allow functions to return void.
Now the callback_manager supports functions that either return a bool
or void. This is useful in plugins that have been using void all along
and in that situation we really do not want one plugin to stop the
calls midway (although there may be cases were we do want that later).

Added a test to verify that the order using priorities works as expected.